+++ serializable_test.ts
@@ -3,7 +3,6 @@
 import { assert, assertEquals, test } from "./test_deps.ts";
 import { Serializable, TransformKey } from "./serializable.ts";
 import { SerializeProperty } from "./serialize_property.ts";
-import { composeStrategy } from "./strategy/compose_strategy.ts";
 
 test({
   name: "adds methods to extended classes",
@@ -20,24 +19,6 @@ test({
     assertEquals(testObj.toJSON(), `{"test":99}`);
     assertEquals(new TestClass().fromJSON({ test: 88 }).test, 88);
     assertEquals(testObj.tsSerialize(), { test: 99 });
-  },
-});
-
-test({
-  name: "composeStrategy composes a List of functions",
-  fn() {
-    const addLetter = (letter: string) => (v: string) => `${v}${letter}`;
-    const shout = (v: string) => `${v}!!!`;
-    const strategy = composeStrategy(
-      addLetter(" "),
-      addLetter("W"),
-      addLetter("o"),
-      addLetter("r"),
-      addLetter("l"),
-      addLetter("d"),
-      shout,
-    );
-    assertEquals(strategy("Hello"), "Hello World!!!");
   },
 });
